Make client configs a lookup, add JS

+++ lib/tasks/generate_client.rake
@@ -2,6 +2,31 @@ require 'open-uri'
 require 'fileutils'
 require_relative '../swagger_codegen'
 
+
+SwaggerLanguageConfigs = {
+
+  "ruby" => lambda do |version|
+    {
+      gemName: 'a15k_interactions_api',
+      gemHomepage: 'https://a15k.org/clients/ruby',
+      gemRequiredRubyVersion: '>= 2.4',
+      moduleName: "A15kInteractions",
+      gemVersion: version,
+    }
+  end,
+
+  "javascript" => lambda do |version|
+    {
+      moduleName: "A15kInteractions",
+      modelPackage: "A15kInteractions",
+      projectName: "A15kInteractionsClient",
+      projectVersion: version,
+      usePromises: true,
+    }
+  end
+
+}
+
 desc <<-DESC.strip_heredoc
   Generate an API client in the app/clients directory.  swagger-codegen
   must be installed.  Run like `rake generate_client[0,ruby]` for the
@@ -17,12 +42,8 @@ task :generate_client, [:api_major_version, :language] => :environment do |tt,ar
     api_exact_version = json[:info][:version]
     output_dir = "#{Rails.application.root}/clients/#{api_exact_version}/#{language}"
 
-    config = case language
-    when "ruby"
-      ruby_config(api_exact_version: api_exact_version)
-    else
-      raise "Don't know #{language} config options yet, see `swagger-codegen config-help -l #{language}"
-    end
+    config = SwaggerLanguageConfigs[language]
+    raise "Don't know #{language} config options yet, see `swagger-codegen config-help -l #{language}" unless config
 
     # Clean out anything that use to be there so old bindings do come back to life
     FileUtils.rm_rf(output_dir)
@@ -30,17 +51,7 @@ task :generate_client, [:api_major_version, :language] => :environment do |tt,ar
     {
       cmd_options: %W[-l #{language}],
       output_dir: output_dir,
-      config: config
+      config: config.call(api_exact_version)
     }
   end
-end
-
-def ruby_config(api_exact_version:)
-  {
-    gemName: 'a15k_interactions_api',
-    gemHomepage: 'https://a15k.org/clients/ruby',
-    gemRequiredRubyVersion: '>= 2.4',
-    moduleName: "A15kInteractions",
-    gemVersion: api_exact_version,
-  }
 end
